Extract butler-touching helpers into butlerQueries module
Move the ten small helpers that need a Butler (or one of its
products like an Exposure or DimensionRecord) out of `utils.py`
into a new `butlerQueries` module:

- getNumExpectedItems — STS-driven detector count fallback for
  partial focal-plane reads
- removeDetector — DataCoordinate filtering used by RedisHelper
- summaryStatsToDict — `ExposureSummaryStats` → JSON-friendly dict
- getAirmass — finite-airmass extractor from an `Exposure`
- getExpIdOrVisitId — exposure-or-visit ID resolver
- getExpRecordFromVisitRecord / getVisitRecordFromExpRecord —
  exposure ↔ visit record conversion via `butler.registry`
- getExpRecordFromId — exposure ID → record lookup
- getCurrentOutputRun — current output collection tip lookup
- getEquivalentDataId — exposure → visit/group data ID rewriter

These are grouped together because they all need a real Butler
(or an object produced by one) to do their job, so they cannot be
exercised in a unit test without either a real Butler or a
substantial mock — that is what the integration suite under
tests/ci/ already covers.

`utils.py` is now down to ~190 lines and contains only setupSentry,
checkRubinTvExternalPackages, catchPrintOutput, DurationResult /
logDuration / timeFunction. Subsequent commits will move the timing
helpers into `timing.py` (where they belong alongside `BoxCarTimer`),
move setupSentry / checkRubinTvExternalPackages into a small
startupChecks module, move catchPrintOutput into its sole consumer
(rubinTv.py), and finally delete the now-empty utils.py.

The previously-unused imports (json, numpy, lsst.daf.butler, etc.)
have been removed from `utils.py`. All callers across the package
have been updated to import from butlerQueries instead of utils.

Verified:
- mypy and flake8 clean across the package, scripts, and tests.
- The full unit test suite (185 tests) passes.
- Every dependent in-package module imports cleanly.

Written by Claude Code 4.6 Opus (high-effort).
